Enable filtering of NADs when not on node
Exposes a filter function that can be propagated to the NAD Controller
by a controller manager. This is used by controller managers to
determine if the network should be ensured or not during NAD handling.

Additional wiring done for adding support for controller manager to call
Reconcile on NAD controller to queue a NAD key when something like pod
tracker or egress IP signal that a node had become active for a network
and needs reconcilitation.

Signed-off-by: Tim Rozet <trozet@nvidia.com>