Allow specifying custom ClusterIP for Services
Introduces new behavior, where if a Headless Service is created with
k8s.ovn.org/cluster-ip annotated with an custom IP address, then a
service will be created using that custom IP address.

The custom IP address may not be in the range of the Kubernetes Service CIDR.

Additionally, the load balancer is not placed on the Gateway Router (GR).
This is intentional to avoid exposing this Custom IP address anywhere
other than inside the SDN. The load balancer is only configured on the
worker switches.

Signed-off-by: Tim Rozet <trozet@nvidia.com>