Integrate the trackers into controller managers and nad controller
Exposes a TrackerController and Tracker interface. Pod and Egress IP
Trackers implement TrackerController. Controller Managers implement
Tracker, as they can query for if a node is serving a NAD.

Controller Managers expose OnNetworkRefChange and Filter functions.
OnNetworkRefChange is a callback registered with the Pod and Egress IP
Trackers to signal when a NAD goes active on a node. Filter function is
registerd with NAD Controller so that NAD Controller can filter out
rendering networks that do not need to be rendered on the local node.

NetworkManager Interface is extended to expose Reoncile and
ForceReconcile. This allows the OnNetworkRefChange callback to trigger
reconciliation of the NAD Controller when a NAD becomes active/inactive
on a local node. When the callback signals something has changed for a
NAD on a node, we attempt to ForceReconcile the NAD.

ForceReconcile with an active network triggers the underlying Network
Controller to indicate that this network should be ensured even if the
NAD has not changed. This will be important later to force the UDN
controller to reconcile. With an inactive network, ForceReconcile is
used to queue the NAD to be deleted after the grace period.

Signed-off-by: Tim Rozet <trozet@nvidia.com>