fix: Remove duplicate - and sometimes incorrect - defaults from --help
For example, ha-election-lease-duration default changed from 60 to 137
recently but the change was not reflected in the help text.

Regardless, the cli parsing library that generates --help message
already documents defaults where applicable, so we don't need to be
explicit about it.

(Included some minor changes to cli options related to "defaults".)

Signed-off-by: Ihar Hrachyshka <ihrachyshka@nvidia.com>