Enable NAD on Node tracking in UDN controllers
Pass ControllerManager as a Tracker into the UDN controllers. This
allows them to query if a node has a NAD on it.

Enable Layer2/3 controllers to filter out adding remote nodes depending
on if the network is active on their node.

During UDN controller reconcile, search for remote nodes that need to be
added or removed.

Also, fix a bug where remote node GR port was not being cleaned up on
node deletion for layer 2 UDNs.

Signed-off-by: Tim Rozet <trozet@nvidia.com>