ovs-node: Delete transient ports on startup
Since 22d2b24b60, ovn-k modifies /etc/default/ configuration for systemd
services included with rhel and ubuntu so that transient ports are
deleted on ovs service startup. This allowed the node to restore
communication through the NIC after reboot.

Once ovnkube starts, it will rewire the NIC back through the ovs bridge,
and move IP configuration (addresses, routes).

The issue is that 22d2b24b60 was assuming that ovs daemons are running
as systemd services, which is not the case with ovs-node daemonset. For
the latter, it's ovnkube.sh that calls ovs-ctl start. This patch extends
the ovnkube.sh script so that we always call ovs-ctl with
--delete-transient-ports when starting ovs services.

Note: in daemonset scenario, ovnkube should probably not attempt to
modify systemd /etc/ files for ovs services. That said, it's probably
harmless that it does so, since errors on missing /etc/ file are
ignored.

Fixes #5596

Signed-off-by: Ihar Hrachyshka <ihrachyshka@nvidia.com>