node: Move network lookup into addServiceRules
Move the call to GetActiveNetworkForNamespace from the various callers
into the addServiceRules function itself.

This change is primarily aimed at de-coupling the network information
lookup from the service rule deletion path. In functions like
DeleteEndpointSlice, delServiceRules is called before addServiceRules.
The previous pattern required fetching NetInfo before both calls, even
though delServiceRules does not require it.

By encapsulating the lookup within addServiceRules, the calling code is
simplified, and the lookup is only performed when it is actually
needed. This cleans up the logic in the service and endpointslice event
handlers.

Unit tests for addServiceRules are also added to ensure correctness
and prevent regressions.

Signed-off-by: Peng Liu <pliu@redhat.com>