STEP2: Allocate && Annotate tunnel-ids for connect router
This commit uses the global TunnelKeysAllocator to allocate
a tunnelID from the global pool. It also annotates this value
onto CNC for it to be used by ovnkube-controller.
The commit also releases the key when CNC is deleted and
during startup ensures we reserved used keys by CNC so that
other controllers or objects don't use the same key as this
should be unique globally.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>