Fix issues with dualstack detection for nodePort with hostNetwork tests
Before, the E2E tests in question would fail on dualstack clusters
because they sent test traffic to each of the node's IPs (IPv4 and
IPv6). However, a) IP addresses were only put in [] based on the cluster
type and not on the type of the node IP itself. Also, the test would try
to connect to both IP address families, even though the service was only
single stack. With this change, the service will be configured as
require dualstack if it has IP addresses in both families.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>