no-overlay: Add cluster CIDR address set for no-overlay
In no-overlay mode with outbound SNAT enabled, pod-to-pod traffic is incorrectly SNATed. This commit introduces a cluster-wide address set containing all node pod subnets. This address set is used to exempt pod-to-pod traffic from SNAT, while still SNATing pod-to-external traffic.

The new address set is created only when the transport is 'no-overlay' and outbound SNAT is enabled.

Signed-off-by: Peng Liu <pliu@redhat.com>