BGP docs: remove references of CUDN route import in non VRF-Lite config
This does not actually work in non VRF-Lite configuration. When the BGP
process attaches to the VRF to establish the BGP session, linux kernel
assigns as source IP the IP of an interface attached to the VRF. In this
case, the only available is management port IP. The session will fail as
the remote peer cannot route back to that source IP.

There are different options but overall we just need to add general
support for host processes attaching to CUDN VRFs.

For now, restrict the documentation to configurations we are actually
testing.

See: https://github.com/ovn-kubernetes/ovn-kubernetes/issues/5580

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>