no-overlay: add kind support for outbound SNAT in no-overlay mode
Allow the --enable-no-overlay flag in kind.sh to accept an optional
'snat-enabled' argument. This configures the OVN setup to enable
outbound-snat in the no-overlay configuration section.

Updates the daemonset script and templates to propagate this setting
to the ovnkube-config ConfigMap.

Also update 2 CI lanes to verify no-overlay mode with SNAT enabled.

Signed-off-by: Peng Liu <pliu@redhat.com>