STEP2: Create patch ports to connect routers together
This commit creates a patch port on a
connected network router and its corresponding
peer port on the connect-router. For layer3
networks, we also create the remote port
corresponding to the local port.

For layer3 networks we have a port for
each network-node pair and for layer2, we
have a global port per network.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>