refactor how BGP being setup with default network advertised
Trying to create a standalone go program to do BGP setup, it can be called alone,
and be called during installation of a KIND cluster with route advertisement enabled.
So the original dependency of demo.sh during BGP (with default network advertised) setup and
dependency of patch for frr-k8s image that is maintained in another repo can be removed.
frr-k8s version is updated to v0.0.21 and frrimage version is updated to 10.4.1

Signed-off-by: Jean Chen <jechen@redhat.com>